Recover from dead node connection attempts
This commit changes all connection attempts to use live peers.

On connection the peer is removed from the peers list and kept only
in live peers. If a connection fails the peer is removed from the
live peers and returned to peers.

This should reduce unnecessary connection attempt against dead
nodes.